Test file
My original dataset was not being recognized in the p5js sketch. So I am trying to copy the stock data given in the mapbox tutorial to see if another geojson file that is not the original works. Which it should. More specifically p5js was having issue identifying the id attribute of the dataset for some reason.